+++ internal/model/product.go
@@ -12,7 +12,7 @@ type Product struct {
 	Category  Category  `json:"category" gorm:"foreignKey:CategoryID;references:ID"`
 	SKU string    `json:"sku" gorm:"unique;not null"`
 	Name string   `json:"name" gorm:"not null"`
-	Price float64  `json:"price" gorm:"not null"`
+	Price int64  `json:"price" gorm:"not null"`
 	Stock int	 `json:"stock" gorm:"not null"`
 	CreatedAt time.Time `json:"created_at" gorm:"autoCreateTime"`
 	UpdatedAt time.Time `json:"updated_at" gorm:"autoUpdateTime"`
